began fixing unit tests
