+++ package.json
@@ -5,7 +5,7 @@
   "main": "server.js",
   "scripts": {
     "start": "node server.js",
-    "test": "set NODE_ENV dev && mocha --timeout 10000"
+    "test": "mocha -r ts-node/register src/test/test.ts"
   },
   "keywords": [
     "taxii",
@@ -23,11 +23,18 @@
     "uuid-validate": "0.0.2"
   },
   "devDependencies": {
+    "@types/chai": "^4.1.3",
+    "@types/chai-http": "^3.0.4",
     "@types/express": "^4.11.1",
+    "@types/mocha": "^5.2.0",
     "@types/mongoose": "^5.0.10",
     "@types/uuid": "^3.4.3",
     "@types/uuid-validate": "0.0.0",
+    "chai": "^4.1.2",
+    "chai-http": "^4.0.0",
     "concurrently": "^3.5.1",
+    "mocha": "^5.1.1",
+    "ts-node": "^6.0.1",
     "tslint": "^5.9.1",
     "typescript": "^2.8.3"
   }
